requirements.txt
amqp==2.5.1
api2pdf==0.0.11
bcrypt==3.1.7
billiard==3.6.1.0
boto3==1.9.234
botocore==1.12.234
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Django==2.2.5
django-cors-headers==3.1.0
django-filter==2.2.0
django-fsm==2.6.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
djangorestframework-fsm==0.1.2
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
jmespath==0.9.4
kombu==4.6.4
more-itertools==7.2.0
mysqlclient==1.4.4
psutil==5.6.3
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-http-client==3.2.1
pytz==2019.2
redis==3.3.8
requests==2.22.0
s3transfer==0.2.1
scout-apm==2.5.0
sendgrid==6.1.0
six==1.12.0
sqlparse==0.3.0
stripe==2.36.2
urllib3==1.25.6
uWSGI==2.0.18
vine==1.3.0
whitenoise==4.1.4
zipp==0.6.0
